% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{addSignatures}
\alias{addSignatures}
\title{Add molecular signatures to MultiAssayExperiment}
\usage{
addSignatures(mae, ..., intersect_rows = TRUE)
}
\arguments{
\item{mae}{MultiAssayExperiment object.}

\item{...}{named experiments to be added to \code{mae}.}

\item{intersect_rows}{logical flag indicating if only common rows across
experiments should be included. Only set to \code{FALSE} if you know what
you are doing.}
}
\value{
MultiAssayExperiment object with new experiments added.
}
\description{
\code{addSignatures} extends \code{mae} by adding to it new experiments.
Rows consistency is ensured by taking an intersection of rows after new
experiments are added.
}
\examples{
TODO

}
