% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{applyOverColumnGroups}
\alias{applyOverColumnGroups}
\title{Apply function over groups of columns}
\usage{
applyOverColumnGroups(mat, groups, f, ...)
}
\arguments{
\item{mat}{a matrix.}

\item{groups}{a vector giving columns grouping.}

\item{f}{function to be applied.}

\item{...}{optional arguments to \code{f}.}
}
\value{
a matrix of dimensions \code{nrow(mat) x nlevels(groups)}.
}
\description{
Returns a array obtained by applying a function to rows of submatrices of the
input matrix, where the submatrices are divided into specified groups of
columns.
}
