% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctss_to_counts_data.table.R
\name{bed_to_anno}
\alias{bed_to_anno}
\title{Read in a BED file to crate "anno" data.table object.}
\usage{
bed_to_anno(bed_file)
}
\arguments{
\item{bed_file}{paths to bed file (BED4 or BED6)}
}
\value{
data.table object with genomic coordinates (6 columns)
}
\description{
Read in a BED file to crate "anno" data.table object.
}
