% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chip_atlas_enhancers}
\alias{chip_atlas_enhancers}
\title{ChIP-Atlas FANTOM5 enhancers intersection matrix}
\format{
A \code{Matrix} with 63285 rows and 15133 columns. Row names corresponds to promoters
names, column names are formatted as TranscriptionFactor_Origin_Cell_ExperimentID
(eg. PARK7_Neural_SH-SY5Y_DRX000550, MLL-AF6_Blood_ML-2_DRX001460).
}
\usage{
chip_atlas_enhancers
}
\description{
An intersection matrix describing overlaps between ChIP-Atlas's ChIP-seq tracks
and \code{\link{enhancers_f5}}. To find overlapping regions promoters were extended
by 500bp in both directions.
}
\keyword{datasets}
