% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chip_atlas_meta}
\alias{chip_atlas_meta}
\title{ChIP-Atlas metadata}
\format{
A \code{data.frame} with 13891 rows and 6 columns.
\describe{
\item{id}{Character giving internal experiment ID.}
\item{tf}{Character giving transcription factor name.}
\item{tf_dbd}{Character giving transcription factor DNA binding domain
family, as per CIS BP database.}
\item{biotype}{Character giving experiment biological origin.}
\item{study}{Character giving study ID.}
}
}
\usage{
chip_atlas_meta
}
\description{
Metadata associated with \code{chip_atlas_promoters} and
\code{chip_atlas_enhancers}.
}
\keyword{datasets}
