% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chip_atlas_symbol}
\alias{chip_atlas_symbol}
\title{ChIP-Atlas gene level interaction matrix}
\format{
A \code{Matrix} with 25925 rows and 15133 columns. Row names corresponds to gene's
symbols, column names are formatted as TranscriptionFactor_Origin_Cell_ExperimentID
(eg. PARK7_Neural_SH-SY5Y_DRX000550, MLL-AF6_Blood_ML-2_DRX001460).
}
\usage{
chip_atlas_symbol
}
\description{
A matrix describing interactions between ChIP-Atlas's ChIP-seq tracks
and human genes (hg38) as defined by \code{\link{promoters}}. The interaction score
for each gene and transcription factor is a \emph{maximum} of transcription factor occurences
among the promoters associated with a given gene. Promoters were assigned to thier target
genes based on gene's symbols.
}
\keyword{datasets}
