% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseInteractionMatrix}
\alias{collapseInteractionMatrix}
\title{collapseInteractionMatrix}
\usage{
collapseInteractionMatrix(
  mat,
  meta,
  cl,
  alpha = 0.5,
  purity_feature = "tf",
  min_pruning_purity = 0.5
)
}
\arguments{
\item{pruning_purity}{if purity > pruning_purity then prune}
}
\description{
collapseInteractionMatrix
}
