% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_ctss_to_bigwig}
\alias{convert_ctss_to_bigwig}
\title{Convers CTSS bed file to  BigWig}
\usage{
convert_ctss_to_bigwig(ctss_file, genomeInfo)
}
\arguments{
\item{ctss_file}{path to ctss.bed file}

\item{genomeInfo}{genome info file created e.g. by "rtracklayer::SeqinfoForUCSCGenome("hg38")"}
}
\value{
The function writes two files, one for each strand, (".plus.bw" and ".minus.bw"), in the same directory as the original file
}
\description{
Convers CTSS bed file to  BigWig
}
