% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{counts_to_activity_scores}
\alias{counts_to_activity_scores}
\title{Calculate Activity Scores for Expressim Matrix}
\usage{
counts_to_activity_scores(
  signature_mat,
  counts,
  min_genes_per_sig = 10,
  prior.count = 4,
  control_samples = NULL
)
}
\arguments{
\item{signature_mat}{sparseMatrix, columns represent signatures and rows represent the genes, 1 means that a gene belongs to a pathway}

\item{counts}{a matrix with expression counts}

\item{min_genes_per_sig}{minimum number of genes per signature, signatures with fewer genes are discarded}

\item{prior.count}{number of counts to be added as an offset in log2 transformation, used by edgeR::cpm()}

\item{control_samples}{vector of indices of control samples to be used as reference, NULL by default}
}
\value{
a matrix of activities, rows represent signatures and columns represent the samples
}
\description{
Calculate Activity Scores for Expressim Matrix
}
