% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_utils.R
\name{create_intersection_matrix}
\alias{create_intersection_matrix}
\title{Create Intersection Matrix
this functions creates a "set intersections in a matrix layout", the format used by UpSetR package to visualize intersections/overlaps bewteen different sets.}
\usage{
create_intersection_matrix(list_of_granges, pre_defined_regions = NULL)
}
\arguments{
\item{list_of_granges}{a list of GRanges containing regions of intereset, names(x) will be used to name the columns of the output matrix}

\item{pre_defined_regions}{a list of predefined regions to be used at rows, if NULL (default), the union of all genomic regions in the list_of_granges will be used.}
}
\value{
intersection matrix
}
\description{
Create Intersection Matrix
this functions creates a "set intersections in a matrix layout", the format used by UpSetR package to visualize intersections/overlaps bewteen different sets.
}
