% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_utils.R
\name{create_intersection_vector}
\alias{create_intersection_vector}
\title{Create Intersection Vector}
\usage{
create_intersection_vector(query, pre_defined_regions)
}
\arguments{
\item{query}{GRanges containing query regions e.g. Transcription Factor Chip-Seq data}

\item{pre_defined_regions}{a set set of reference regions like known promoters or enhancers}
}
\value{
a vector of the same lenght as pre_defined_regions, 0 if no overlap, 1 if there is overlap with query regions
}
\description{
Create Intersection Vector
}
