% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctss_to_counts_data.table.R
\name{ctss_data_to_counts}
\alias{ctss_data_to_counts}
\title{Create count table by overlapping CTSS data with genomic coordinates (data.table)}
\usage{
ctss_data_to_counts(ctss_data, anno)
}
\arguments{
\item{ctss_data}{data.table object with CTSS data, created by "read_in_ctss" function}

\item{anno}{data.table object with genomic coordinates, created by "bed_to_anno" function}
}
\value{
count table with number of reads mapping to genomic regions
}
\description{
Create count table by overlapping CTSS data with genomic coordinates (data.table)
}
