% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctss_to_counts_data.table.R
\name{ctss_to_counts}
\alias{ctss_to_counts}
\title{CTSS files to count table wrapper}
\usage{
ctss_to_counts(ctss_files, regions_bed)
}
\arguments{
\item{ctss_files}{a character vector with paths to CTSS files, names(ctss_files) should represent sample names}

\item{regions_bed}{path to bed file with genomic regions of interest}
}
\value{
(expression) count table of number of reads within provided genomic regions
}
\description{
CTSS files to count table wrapper
}
