% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{de_res_to_activity_scores}
\alias{de_res_to_activity_scores}
\title{Calculate Activity Scores (as Mean FC of Singature Features)}
\usage{
de_res_to_activity_scores(signature_mat, de_res, min_genes_per_sig = 10)
}
\arguments{
\item{signature_mat}{a sparseMatrix, columns represent signatures and rows represent the genes, 1 means that a gene belongs to a pathway}

\item{de_res}{list of data.frame objects, e.g. DE results}

\item{min_genes_per_sig}{minimum number of genes per signature, signatures with fewer genes are discarded}
}
\value{
a matrix of activities, rows represent signatures and columns represent the DE comparisons
}
\description{
Calculate Activity Scores (as Mean FC of Singature Features)
}
