% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{deseq_DE_wrapper}
\alias{deseq_DE_wrapper}
\title{DeSeq2 wrapper}
\usage{
deseq_DE_wrapper(
  counts,
  sample_annot,
  adj_pv_threshold = 0.05,
  log2fc_threshold = 1
)
}
\arguments{
\item{counts}{data.matrix with the counts data, rownames should represent features and colnames sample names}

\item{sample_annot}{data.frame with sample annotation sample_annot$sample_names should be same as colnames(counts), it has to contain exp_fac column with experimental factor, control needs to be a the first level of the factor (see relevel() )}

\item{adj_pv_threshold}{adjusted p-value significance threshold for DE testing, 0.05 by default}

\item{log2fc_threshold}{log2 Fold Change threshold for DE testing, 1 by default}
}
\value{
a list of differential expression results, each element of the list corresponds to one comparison: each experimental factor level vs control (first factor level)
}
\description{
DeSeq2 wrapper
}
