% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{design2factor}
\alias{design2factor}
\title{Transform design matrix to factor}
\usage{
design2factor(design)
}
\arguments{
\item{design}{design matrix}
}
\value{
factor
}
\description{
Transform design matrix to factor
}
\examples{
design <- matrix(data = c(1, 1, 0, 0, 0, 0, 1, 1),
                 nrow = 4,
                 ncol = 2,
                 dimnames = list(c(paste("sample", 1:4)), c("gr1", "gr2")))
design2factor(design)

}
