% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{edgeR_DE_wrapper}
\alias{edgeR_DE_wrapper}
\title{edgeR wrapper}
\usage{
edgeR_DE_wrapper(
  counts,
  sample_annot,
  design,
  contrasts,
  robust = TRUE,
  plots = TRUE,
  log2fc_threshold = 1,
  adj_pv_threshold = 0.05
)
}
\arguments{
\item{counts}{data.matrix with the counts data, rownames should represent features and colnames sample names}

\item{sample_annot}{data.frame with sample annotation sample_annot$sample_names should be same as colnames(counts), it has to contain exp_fac column with experimental factor, control needs to be a the first level of the factor (see relevel() )}

\item{design}{design matrix}

\item{contrasts}{matrix with contrasts}

\item{robust}{should robust statitsital modeling be used?}

\item{plots}{should plotBCV() and plotQLDisp() be used to generate plots?}

\item{log2fc_threshold}{log2 Fold Change threshold for DE testing, 1 by default#'}

\item{adj_pv_threshold}{adjusted p-value significance threshold for DE testing, 0.05 by default}
}
\value{
a list of differential expression results, each element of the list corresponds to one row of contrast matrix
}
\description{
edgeR wrapper
}
