% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enhancers}
\alias{enhancers}
\title{Enahncers GenomicRanges object}
\format{
A \code{GenomicRanges} object of length 63285, with 4 metadata columns:
\describe{
\item{name}{Feature name.}
\item{hg38_enhancer}{FANTOM5's hg38 enhancer name.}
\item{ep300}{Variable indicating overlapp with EP300 ReMap2020 track.}
\item{repeat_dfam}{variable indicating overlapp with Dfam dyadic track.}
}
}
\usage{
enhancers
}
\description{
A dataset containing FANTOM5's hg38 enhancers overlapped with EP300 peaks
from ReMap2020 and dyadic regions from Dfam database.
}
\keyword{datasets}
