% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enhancers_f5}
\alias{enhancers_f5}
\title{Enahncers GenomicRanges object}
\format{
A \code{GenomicRanges} object of length 63285, with 4 metadata columns:
\describe{
\item{name}{Feature name.}
\item{hg38_enhancer}{FANTOM5's hg38 enhancer name.}
\item{ep300}{Variable indicating overlapp with EP300 ReMap2020 track.}
\item{repeat_dfam}{Variable indicating overlap with Dfam dyadic track.}
\item{tau}{Numeric Tau tissue specificity score calculated based on FANTOM5
expression data.}
\item{encode_blacklist}{Logical indicating overlap with ENCODE black
listed regions.}
\item{ensembl_sm}{Logical indicating overlap with ENSEMBL soft masked
regions.}
}
}
\usage{
enhancers_f5
}
\description{
A dataset containing FANTOM5's hg38 enhancers overlapped with EP300 peaks
from ReMap2020 and dyadic regions from Dfam database.
}
\keyword{datasets}
