% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{extract_a_column}
\alias{extract_a_column}
\title{Extract a column of interest from list of data.frame's}
\usage{
extract_a_column(de_res, column)
}
\arguments{
\item{de_res}{list of data.frame objects, e.g. DE results}

\item{column}{name of the column to extract, e.g. "log2FoldChange"}
}
\value{
extracted column cbind'ed into a matrix
}
\description{
Extract a column of interest from list of data.frame's
}
