% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{fisherTestSignature}
\alias{fisherTestSignature}
\title{Fisher's test signatures}
\usage{
fisherTestSignature(exprss_mat, sign_mat, alternative = "greater")
}
\arguments{
\item{exprss_mat}{a binary expression matrix.}

\item{sign_mat}{a binary signature matrix.}

\item{alternative}{indicates the alternative hypothesis and must be
    one of \code{"two.sided"}, \code{"greater"} or \code{"less"}.
    You can specify just the initial letter.  Only used in the
    \eqn{2 \times 2}{2 by 2} case.}
}
\value{
a list with \code{"p.value"} matrix of p-values and \code{"odds"}
matrix of odds ratios from the Fisher test.
}
\description{
Perform Fisher test between binary matrices columns. The function can be work
using parallelization  if  a parallel backend is registered (eg. \code{doMC}).
}
