% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_genomic_regions.R
\name{gencode_nearest_promoter_both_strands}
\alias{gencode_nearest_promoter_both_strands}
\title{Annotate with Nearest Promoter for Both Strands}
\usage{
gencode_nearest_promoter_both_strands(regions, gencode)
}
\arguments{
\item{regions}{query genomic regions as GRanges object}

\item{gencode}{complete Gencode annotation as GRanges object, e.g. obrained by rtracklayer::import.gff(con = "gencode.v31.annotation.gff3.gz") ,  (gencode\link{ gencode$type == "transcript"} is also OK)}
}
\value{
GRanges with new columns added to the feature metadata: mcols(regions)
}
\description{
The wrappers call the gencode_one_direction_distanceToNearest twice after setting the strand to + and -.
}
