% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_genomic_regions.R
\name{gencode_nearest_promoter_same_strand}
\alias{gencode_nearest_promoter_same_strand}
\title{Annotate with Nearest Promoter on the Same Strand}
\usage{
gencode_nearest_promoter_same_strand(regions, gencode, cut_off_distance = 500)
}
\arguments{
\item{regions}{query genomic regions as GRanges object}

\item{gencode}{complete Gencode annotation as GRanges object, e.g. obrained by rtracklayer::import.gff(con = "gencode.v31.annotation.gff3.gz") ,  (gencode\link{ gencode$type == "transcript"} is also OK)}
}
\value{
GRanges with new columns added to the feature metadata: mcols(regions)
}
\description{
Annotate with Nearest Promoter on the Same Strand
}
