% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{gencode_one_direction_distanceToNearest}
\alias{gencode_one_direction_distanceToNearest}
\title{Annotation Worker by Same Strand DistanceToNearest}
\usage{
gencode_one_direction_distanceToNearest(regions, gencode)
}
\arguments{
\item{regions}{query genomic regions as GRanges object}

\item{gencode}{Gencode annotation as GRanges object, usually of one type like gencode\link{ gencode$type == "exon"}}

\item{annotation_label}{label (character) to mark what annotation was used, e.g. "gene", "exon" or "promoters"}
}
\value{
GRanges with new columns added to the feature metadata: mcols(regions)
}
\description{
Annotation Worker by Same Strand DistanceToNearest
}
