% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_genomic_regions.R
\name{gencode_one_direction_findOverlaps}
\alias{gencode_one_direction_findOverlaps}
\title{Annotation Worker by Same Strand Overlap}
\usage{
gencode_one_direction_findOverlaps(regions, gencode, annotation_label = " ")
}
\arguments{
\item{regions}{query genomic regions as GRanges object}

\item{gencode}{Gencode annotation as GRanges object, usually of one type like gencode$type == "exon"}

\item{annotation_label}{label (character) to mark what annotation was used, e.g. "gene", "exon" or "promoters"}
}
\value{
GRanges with new columns added to the feature metadata: mcols(regions)
}
\description{
Usually used by the gencode_one_direction_annotator function.
}
