% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_genomic_regions.R
\name{gencode_strandless_annotator}
\alias{gencode_strandless_annotator}
\title{Annotating Strandless regions by Overlap}
\usage{
gencode_strandless_annotator(regions, gencode, output_option = 2)
}
\arguments{
\item{regions}{query genomic regions as GRanges object}

\item{gencode}{complete Gencode annotation as GRanges object, e.g. obrained by rtracklayer::import.gff(con = "~/projects/resources/gencode/gencode31/gencode.v31.annotation.gff3.gz")}
}
\value{
GRanges with new columns added to the feature metadata: mcols(regions)
}
\description{
This is mostly for bidirectional/strandless enhancer regions
The wrappers call the gencode_one_direction_annotator twice after setting the strand to + and -.
Finally the output is nicely summarized
}
