% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_genomic_regions.R
\name{gencode_two_direction_annotator}
\alias{gencode_two_direction_annotator}
\title{Annotating Both Strands by Overlap}
\usage{
gencode_two_direction_annotator(regions, gencode)
}
\arguments{
\item{regions}{query genomic regions as GRanges object}

\item{gencode}{complete Gencode annotation as GRanges object, e.g. obrained by rtracklayer::import.gff(con = "~/projects/resources/gencode/gencode31/gencode.v31.annotation.gff3.gz")}
}
\value{
GRanges with new columns added to the feature metadata: mcols(regions)
}
\description{
A wrapper that calls gencode_one_direction_annotator on same and inverted strand
to annotate overlap on both strands
}
