% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{get_fc_matrix}
\alias{get_fc_matrix}
\title{Get Matrix of Fold Changes (FCs)}
\usage{
get_fc_matrix(de_res)
}
\arguments{
\item{de_res}{a list of DE results data frame with log2FoldChange column}
}
\value{
matrix of fold changes
}
\description{
Get Matrix of Fold Changes (FCs)
}
