% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getAvgCoeff}
\alias{getAvgCoeff}
\title{Calculate average coefficients matrix}
\usage{
getAvgCoeff(models, lambda = "lambda.min", drop_intercept = TRUE)
}
\arguments{
\item{models}{list of \code{cv.glmnet} objects.}

\item{lambda}{string indicating which lambda to use.}

\item{drop_intercept}{logical indicating if intercept should be droped from
the output.}

\item{a}{average coefficients matrix}
}
\description{
Calculate average coefficients matrix
}
