% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getVarianceWeightedAvgCoeff}
\alias{getVarianceWeightedAvgCoeff}
\title{Calculate variance weighted average coefficients matrix}
\usage{
getVarianceWeightedAvgCoeff(pvalues, groups)
}
\arguments{
\item{pvalues}{list of \code{data.frame}s outputs from \code{ridgePvals}.}

\item{groups}{factor giving the grouping.}
}
\value{
variance weighted average coefficients matrix
}
\description{
Calculate variance weighted average coefficients matrix
}
