% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctss_to_counts_data.table.R
\name{granges_to_anno}
\alias{granges_to_anno}
\title{Convert GRanges object to "anno" data.table object}
\usage{
granges_to_anno(granges)
}
\arguments{
\item{granges}{a GRanges object to be converted to data.table}
}
\value{
data.table object with genomic coordinates (6 columns)
}
\description{
Convert GRanges object to "anno" data.table object
}
