% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intersectGR}
\alias{intersectGR}
\title{Intersect GRanges objects}
\usage{
intersectGR(a, b, ...)
}
\arguments{
\item{a}{GRanges object.}

\item{b}{GRanges object.}

\item{...}{other arguments internally passed to
\code{\link[GenomicRanges]{findOverlaps}}.}
}
\value{
GRanges object which is a subset of \code{a}.
}
\description{
\code{intersectGR} intersect two GRanges objects and return ranges from
\code{a} that were overlapped by ranges in \code{b}.
}
