% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isTRUEorFALSE}
\alias{isTRUEorFALSE}
\title{Check if argument is a binary flag}
\usage{
isTRUEorFALSE(x)
}
\arguments{
\item{x}{object to test}
}
\value{
binary flag
}
\description{
Check if argument is a binary flag
}
