% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{limma_DE_wrapper}
\alias{limma_DE_wrapper}
\title{limma differential analysis wrapper}
\usage{
limma_DE_wrapper(
  data,
  design,
  contrasts,
  log2fc_threshold = 1,
  adj_pv_threshold = 0.05
)
}
\arguments{
\item{data}{a matrix of expression values or activities, data should be normally distributed (e.g. normalized and log2 transformed)}

\item{design}{design matrix}

\item{contrasts}{matrix with contrasts}

\item{log2fc_threshold}{log2 Fold Change threshold for DE testing, 1 by default#'}

\item{adj_pv_threshold}{adjusted p-value significance threshold for DE testing, 0.05 by default}
}
\value{
a list of differential expression results, each element of the list corresponds to one row of contrast matrix
}
\description{
limma differential analysis wrapper
}
