% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctss_to_counts_data.table.R
\name{list_the_ctss_files}
\alias{list_the_ctss_files}
\title{List the CTSS Files}
\usage{
list_the_ctss_files(path)
}
\arguments{
\item{path}{path to the directory containing the CTSS files}
}
\value{
a character vector of bam files (full path), with "names" representing shortened file names
}
\description{
List the CTSS Files
}
