% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{list_to_matrix}
\alias{list_to_matrix}
\title{Create a 0,1 encoded matrix from a list of signatures}
\usage{
list_to_matrix(sig_list)
}
\arguments{
\item{sig_list}{a list gene signatures}
}
\value{
0,1 encoded matrix with columns representing signatures and rows representing the genes, 1 means that a gene belongs to a pathway
}
\description{
Create a 0,1 encoded matrix from a list of signatures
}
