% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{list_to_sparse_matrix}
\alias{list_to_sparse_matrix}
\title{Create a sparseMatrix from a list of signatures}
\usage{
list_to_sparse_matrix(sig_list)
}
\arguments{
\item{sig_list}{a list gene signatures}
}
\value{
a sparseMatrix, columns represent signatures and rows represent the genes, TRUE means that a gene belongs to a pathway
}
\description{
Create a sparseMatrix from a list of signatures
}
