% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mae}
\alias{mae}
\title{Calculate Mean Absolute Error}
\usage{
mae(y, yhat, ...)
}
\arguments{
\item{y}{numeric vector of observed expression values.}

\item{yhat}{numeric vector of predicted expression values.}

\item{..}{not used.}
}
\value{
numeric vector
}
\description{
Calculate Mean Absolute Error
}
