% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{maeSummary}
\alias{maeSummary}
\title{Helper summarizing MAE object}
\usage{
maeSummary(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object.}
}
\value{
named list giving number of rows and columns, overall mean and
standard deviation in \code{mae}'s experiments.
}
\description{
Helper summarizing MAE object
}
