% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_balanced_colorscale}
\alias{make_balanced_colorscale}
\title{Make Colorscale ballanced around zero
Wrapper/Convenience function to gplots::colorpanel()
This function generates pallette of colors for a heatmap where "mid" color is centered on 0,
e.g. to separate the up-regulated/positive values ( "high") and down-regulated/negative values ("low")}
\usage{
make_balanced_colorscale(
  dat_mat,
  col_res_number = 10,
  low = "steelblue",
  mid = "white",
  high = "red"
)
}
\arguments{
\item{dat_mat}{data matrix that will be used for ploting, should contain positive and negative values}

\item{col_res_number}{number of color elements in the panel/pallette, passed on to gplots::colorpanel}

\item{low}{optional, default "steelblue"}

\item{mid}{optional, default "white"}

\item{high}{optional, default "red"}
}
\value{
Vector of HTML-style RGB colors
}
\description{
Make Colorscale ballanced around zero
Wrapper/Convenience function to gplots::colorpanel()
This function generates pallette of colors for a heatmap where "mid" color is centered on 0,
e.g. to separate the up-regulated/positive values ( "high") and down-regulated/negative values ("low")
}
