% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{make_enrichment_score_matrix}
\alias{make_enrichment_score_matrix}
\title{Make Enrichment Score Matrix}
\usage{
make_enrichment_score_matrix(fgseaRes, score = c("NES", "ES"))
}
\arguments{
\item{fgseaRes}{list of fgsea() results as generated by run_fgsea()}

\item{score}{choose between two enrichment scores: "NES" or "ES"}
}
\value{
a matrix with enrichment scores ("NES" or "ES"), each column represents DE comparison, each row represents a pathway/signature
}
\description{
Make Enrichment Score Matrix
}
