% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeClusterName}
\alias{makeClusterName}
\title{Make cluster name based on metadata}
\usage{
makeClusterName(
  meta,
  columns = c("tf", "biotype", "study", "tf_dbd"),
  mixed_th = 0.5
)
}
\description{
Make cluster name based on metadata
}
