% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeClusterName2}
\alias{makeClusterName2}
\title{Make cluster name based on metadata ver. 2}
\usage{
makeClusterName2(
  meta,
  columns = c("tf", "biotype", "study", "tf_dbd"),
  inclusion_th = 0.25,
  mixed_th = 0.5
)
}
\description{
Make cluster name based on metadata ver. 2
}
