% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctss_to_counts_data.table.R
\name{mapping_stats}
\alias{mapping_stats}
\title{Mapping statistics for CAGE data from CTSS files (for Quality Control)}
\usage{
mapping_stats(ctss_data, simple_anno)
}
\arguments{
\item{ctss_data}{data.table object with CTSS data, created by "read_in_ctss" function}

\item{simple_anno}{data.table object with genomic coordinates, data( simple_anno, package = "xcore")}
}
\value{
(map_stats) table of number of reads within provided genomic regions: promoters, introns, exons, intergenic
}
\description{
Mapping statistics for CAGE data from CTSS files (for Quality Control)
}
