% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{matrix_to_list}
\alias{matrix_to_list}
\title{Create a list of signatures from 0,1 encoded matrix}
\usage{
matrix_to_list(int_mat)
}
\arguments{
\item{int_mat}{0,1 encoded matrix with columns representing signatures and rows representing the genes, 1 means that a gene belongs to a pathway}
}
\value{
a list gene signatures
}
\description{
Create a list of signatures from 0,1 encoded matrix
}
