% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_utils.R
\name{merge_list_of_granges}
\alias{merge_list_of_granges}
\title{Merge list of GRanges}
\usage{
merge_list_of_granges(list_of_granges)
}
\arguments{
\item{list_of_granges}{a list, each element should be a GRanges object}
}
\value{
a GRanges object with all the regions merged
}
\description{
Merge list of GRanges
}
