% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{modelGeneExpression}
\alias{modelGeneExpression}
\title{Gene expression modeling pipeline}
\usage{
modelGeneExpression(
  mae,
  yname,
  uname,
  xnames,
  design,
  standardize = TRUE,
  parallel = FALSE,
  pvalues = TRUE,
  average_replicates = TRUE,
  precalcmodels = NULL,
  ...
)
}
\arguments{
\item{mae}{MultiAssayExperiment object.# TODO X has to be filtered at least for zeros  and ones !!!}

\item{yname}{String response matrix name.}

\item{uname}{String response offset name.}

\item{xnames}{Character signatures names.}

\item{standardize}{Logical should signatures be scaled? scale(X).}

\item{parallel}{parallel argument to glmnet::cv.glmnet. Advised to be set to FALSE. Lower number of folds could be used for speed up.}

\item{...}{arguments passed to glmnet::cv.glmnet.}
}
\value{
foo
}
\description{
foo bar
}
\examples{
TODO

}
