% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mse}
\alias{mse}
\title{Calculate Mean Squared Error}
\usage{
mse(y, yhat, ...)
}
\arguments{
\item{y}{numeric vector of observed expression values.}

\item{yhat}{numeric vector of predicted expression values.}

\item{offset}{numeric vector giving basal expression level.}
}
\value{
numeric vector
}
\description{
Calculate Mean Squared Error
}
