% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_utils.R
\name{name_from_granges}
\alias{name_from_granges}
\title{Name the regions based on genomic coordinates}
\usage{
name_from_granges(granges)
}
\arguments{
\item{granges}{a GRanges object}
}
\value{
a character vector of coordinate-based names
}
\description{
Name the regions based on genomic coordinates
}
