% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize the count table with edgeR}
\usage{
normalize_counts(counts, method = "RLE", log = FALSE, prior.count = 1, ...)
}
\arguments{
\item{counts}{expression table with counts}

\item{method}{passed to to edgeR::calcNormFactors see ?edgeR::calcNormFactors}

\item{log}{passed on to edgeR::cpm , should the normalized counts be log2 transformed (default FALSE)}

\item{prior.count}{average count to be added to each observation to avoid taking log of zero. Used only if log=TRUE.}

\item{...}{other parameter passed on to edgeR::calcNormFactors or edgeR::cpm}
}
\value{
normalized count table
}
\description{
Normalize the count table with edgeR
}
