% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_by_variance}
\alias{order_by_variance}
\title{Simple wrapper to order data by variance}
\usage{
order_by_variance(x)
}
\arguments{
\item{x}{table to be ordered}
}
\value{
table with most variant features (rows) on top
}
\description{
Simple wrapper to order data by variance
}
