% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctss_to_counts_data.table.R
\name{plot_mapping_stats}
\alias{plot_mapping_stats}
\title{Plot Mapping Statistics}
\usage{
plot_mapping_stats(qc_stats)
}
\arguments{
\item{qc_stats}{table with QC stats generated by mapping_stats() function}
}
\value{
a QC mapping plot as ggplot object
}
\description{
Plot Mapping Statistics
}
