% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotExperimentsHeatmap}
\alias{plotExperimentsHeatmap}
\title{Experiment's heatmap}
\usage{
plotExperimentsHeatmap(
  mat,
  meta,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  angle_col = 45,
  treeheight_col = 0,
  fontsize = 8,
  fontsize_row = 7,
  fontsize_col = 7,
  ...
)
}
\arguments{
\item{mat}{Numeric matrix of pairwise correlations to be plotted.}

\item{meta}{data.table metadata associated with \code{mat}. Should contain
\code{id} column, which is used to subset \code{mat}. Other character
columns are used as heatmap annotations.}

\item{cluster_rows}{boolean values determining if rows should be clustered or \code{hclust} object,}

\item{cluster_cols}{boolean values determining if columns should be clustered or \code{hclust} object.}

\item{angle_col}{angle of the column labels, right now one can choose only from few 
predefined options (0, 45, 90, 270 and 315)}

\item{treeheight_col}{the height of a tree for columns, if these are clustered. 
Default value 50 points.}

\item{fontsize}{base fontsize for the plot}

\item{fontsize_row}{fontsize for rownames (Default: fontsize)}

\item{fontsize_col}{fontsize for colnames (Default: fontsize)}

\item{...}{arguments passed to \code{\link[pheatmap]{pheatmap}} function.}
}
\value{
Invisibly a pheatmap object. See \code{\link[pheatmap]{pheatmap}} for
more informations.
}
\description{
\code{plotExperimentsHeatmap} crate a heatmap for a group o experiments
(columns) based on interaction matrix. As function does not use sparsity
aware methods for distance calculation it is advised to not use too large
matrices. \code{\link[colorspace]{qualitative_hcl}} color palette is used
to allow higer number of categories in the annotation. No scaling is applied
to \code{mat}!
}
