% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Heatmap}
\usage{
plotHeatmap(
  x,
  breaks,
  col = rdBuPalette(length(breaks) - 1),
  rownames = TRUE,
  colnames = TRUE,
  clust_rows = TRUE,
  na.color = "grey",
  mar = c(6.1, 0.1, 4.1, 22.1),
  cex = 1,
  cex.main = 0.9,
  cex.axis = 0.8,
  ...
)
}
\description{
Heatmap
}
