% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotHeatmapScaleBar}
\alias{plotHeatmapScaleBar}
\title{Scale bar to use with heatmaps}
\usage{
plotHeatmapScaleBar(
  breaks,
  col = rdBuPalette(length(breaks) - 1),
  side = 2,
  nticks = 3,
  mar = c(2.1, 3.1, 4.1, 1.1),
  cex = 0.75,
  cex.main = 0.9,
  ...
)
}
\description{
Scale bar to use with heatmaps
}
