% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Histogram with summary}
\usage{
plotHistogram(x, breaks = 40, cex = 0.7, xlab = "", ...)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired.}

\item{breaks}{one of:
    \itemize{
      \item a vector giving the breakpoints between histogram cells,
      \item a function to compute the vector of breakpoints,
      \item a single number giving the number of cells for the histogram,
      \item a character string naming an algorithm to compute the
      number of cells (see \sQuote{Details}),
      \item a function to compute the number of cells.
    }
    In the last three cases the number is a suggestion only; as the
    breakpoints will be set to \code{\link{pretty}} values, the number
    is limited to \code{1e6} (with a warning if it was larger).  If
    \code{breaks} is a function, the \code{x} vector is supplied to it
    as the only argument (and the number of breaks is only limited by
    the amount of available memory).
  }

\item{xlab}{main title and axis labels: these arguments to
    \code{\link[graphics]{title}()} get \dQuote{smart} defaults here, e.g., the default
    \code{ylab} is \code{"Frequency"} iff \code{freq} is true.}

\item{...}{further arguments and \link[graphics]{graphical parameters} passed to
    \code{\link[graphics]{plot.histogram}} and thence to \code{\link[graphics]{title}} and
    \code{\link[graphics]{axis}} (if \code{plot = TRUE}).}
}
\value{
an object of class "histogram".
}
\description{
\code{plotHistogram} create a histogram with summary on the upper margin.
Refer to \code{\link[graphics]{hist}} documentation for more informations.
}
