% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{poolSE}
\alias{poolSE}
\title{Pool Standard Error / Standard Deviation}
\usage{
poolSE(x)
}
\description{
Cohen, J. (1988) TODO check out https://www.statisticshowto.com/find-pooled-sample-standard-error/, https://www.statisticshowto.com/pooled-standard-deviation/
}
