% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepareCountsForRegression}
\alias{prepareCountsForRegression}
\title{Process count matrix for modeling}
\usage{
prepareCountsForRegression(
  mat,
  base_idx,
  group = NULL,
  log2 = TRUE,
  pseudo_count = 1
)
}
\description{
Process count matrix for modeling
}
