% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promoters}
\alias{promoters}
\title{Promoters GenomicRanges object}
\format{
A GenomicRanges object of length 209911, with 11 metadata columns:
\describe{
\item{name}{Promotor name.}
\item{score}{Numeric vector.}
\item{SYMBOL_F5_annot}{Gene symbol of associated gene as defined by FANTOM5 annotation.}
\item{ENTREZID_F5_annot}{ENTREZ ID of associated gene as defined by FANTOM5 annotation.}
\item{SYMBOL_gencode}{Gene symbol of associated gene as defined by GENCODE annotation.}
\item{ENTREZID_gencode}{ENTREZ ID of associated gene as defined by GENCODE annotation.}
\item{gene_type_gencode}{Gene type of associated gene as defined by GENCODE annotation.}
\item{SYMBOL_ucsc}{Gene symbol of associated gene as defined by UCSC annotation.}
\item{ENTREZID_ucsc}{ENTREZ ID of associated gene as defined by UCSC annotation.}
\item{ENTREZID}{ENTREZ ID of associated gene as defined by all three annotation sources.
Prevalence of annotations: UCSC > GENCODE > FANTOM5.}
\item{SYMBOL}{Gene symbol of associated gene as defined by all three annotation sources.
Prevalence of annotations: UCSC > GENCODE > FANTOM5.}
}
}
\usage{
promoters
}
\description{
A dataset containing FANTOM5's hg38 promoters with FANTOM5's annotations.
Additionally this dataset was also annotated to nearest features in GENCODE
ver. 38 annotation and UCSC hg38 knownGene annotation ver. 3.13.0.
}
\keyword{datasets}
