% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promoters_detailed}
\alias{promoters_detailed}
\title{Detailed promoters GenomicRanges object}
\format{
A GenomicRanges object of length 209911, with 24 metadata columns:
\describe{
\item{name}{Promotor name.}
\item{score}{Numeric vector.}
\item{SYMBOL_F5_annot}{Gene symbol of associated gene as defined by FANTOM5 annotation.}
\item{ENTREZID_F5_annot}{ENTREZ ID of associated gene as defined by FANTOM5 annotation.}
\item{distance_F5_annot}{Distance to associated gene as defined by FANTOM5 annotation.}
\item{SYMBOL_gencode}{Gene symbol of associated gene as defined by GENCODE annotation.}
\item{ENTREZID_gencode}{ENTREZ ID of associated gene as defined by GENCODE annotation.}
\item{gene_type_gencode}{Gene type of associated gene as defined by GENCODE annotation.}
\item{distance_gencode}{Distance to associated gene as defined by GENCODE annotation.}
\item{SYMBOL_ucsc}{Gene symbol of associated gene as defined by UCSC annotation.}
\item{ENTREZID_ucsc}{ENTREZ ID of associated gene as defined by UCSC annotation.}
\item{distance_ucsc}{Distance to associated gene as defined by UCSC annotation.}
\item{ENTREZID}{ENTREZ ID of associated gene as defined by all three annotation sources.
Prevalence of annotations: UCSC > GENCODE > FANTOM5.}
\item{SYMBOL}{Gene symbol of associated gene as defined by all three annotation sources.
Prevalence of annotations: UCSC > GENCODE > FANTOM5.}
\item{ep300}{variable indicating overlapp with EP300 ReMap2020 track.}
\item{enhancer}{variable indicating overlapp with \code{\link{enhancers}}.}
\item{repeat_dfam}{variable indicating overlapp with Dfam dyadic track.}
\item{annotation}{Feature annotation type as defined by overlapp with GENCODE annotation.}
\item{symbol}{Gene symbol of associated gene as defined by overlapp with GENCODE annotation.}
\item{gene_type}{Gene type of associated gene as defined by overlapp with GENCODE annotation.}
\item{level}{Character vector.}
\item{opposite_strand_annotation}{Feature annotation type as defined by opposite strand
overlapp with GENCODE annotation.}
\item{opposite_strand_symbol}{Gene symbol of associated gene as defined by opposite strand
overlapp with GENCODE annotation.}
\item{opposite_gene_type}{Gene type of associated gene as defined by opposite strand overlapp
with GENCODE annotation.}
}
}
\usage{
promoters_detailed
}
\description{
A dataset containing FANTOM5's hg38 promoters with FANTOM5's annotations.
Additionally this dataset was also annotated to nearest features in GENCODE
ver. 38 annotation and UCSC hg38 knownGene annotation ver. 3.13.0.
Further object was overlapped with EP300 peaks from ReMap2020, \code{\link{enhancers}} object
and dyadic regions from Dfam database. Moreover the dataset is overlapped with GENCODE annotation
on both strands.
}
\keyword{datasets}
