% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pruneClusterMeta}
\alias{pruneClusterMeta}
\title{Prune metadata}
\usage{
pruneClusterMeta(meta, col)
}
\arguments{
\item{col}{pruning column}
}
\description{
Prunes metadata based on selected column such that only entries with most
frequent items remains.
}
