% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{rank_genes}
\alias{rank_genes}
\title{Rank Genes for fgsea}
\usage{
rank_genes(de_results, based_on = c("pvalue", "log2FoldChange"))
}
\arguments{
\item{de_results}{data.frame with DE results from edgeR or DeSeq, should contail "pvalue" and "log2FoldChange" columns and DE stats for all features}

\item{based_on}{choose if "pvalue" or "log2FoldChange" should be used to rank the genes/featrues}
}
\value{
sorted vector of ranks with names that correspond to gene names (required input to fgse() function)
}
\description{
Rank Genes for fgsea
}
