% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{ranks_from_FC_table}
\alias{ranks_from_FC_table}
\title{Create list of ranks from table of FoldChanges}
\usage{
ranks_from_FC_table(fc_table)
}
\arguments{
\item{fc_table}{a table of FoldChanges}
}
\value{
list of rank vectors, each rank vector can be used as input to fgsea() function
}
\description{
Create list of ranks from table of FoldChanges
}
