% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctss_to_counts_data.table.R
\name{read_in_ctss}
\alias{read_in_ctss}
\title{Read in CTSS files to data.table object}
\usage{
read_in_ctss(ctss_files)
}
\arguments{
\item{ctss_files}{character vector with the paths to CTSS files, names(ctss_files) should contain the sample names}
}
\value{
data.table object with all the data from CTSS files
}
\description{
Read in CTSS files to data.table object
}
