% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{read_msigdb_gmts}
\alias{read_msigdb_gmts}
\title{Read Collections of MSigDB signatures}
\usage{
read_msigdb_gmts(path_to_gmts, msigdb_version = ".v7.0.entrez.gmt")
}
\arguments{
\item{path_to_gmts}{a path to directory containing .gmt files downloaded from http://software.broadinstitute.org/gsea/downloads.jsp#msigdb}

\item{msigdb_version}{a version of MSigDB to read in, e.g. ".v7.0.entrez.gmt"}
}
\value{
a list, each element of the list represents one signature, e.g. c2.cgp___RUNNE_GENDER_EFFECT_UP" represent "RUNNE_GENDER_EFFECT_UP" signature from "c2.cgp" collection
}
\description{
Read Collections of MSigDB signatures
}
