% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remap_enhancers}
\alias{remap_enhancers}
\title{ReMap2020 and FANTOM5 enhancers intersection matrix}
\format{
A Matrix with 63285 rows and 5728 columns. Row names corresponds to enhancers
names, column names are formatted as ExperimentID.TranscriptionFactor.Biotype.
}
\usage{
remap_enhancers
}
\description{
An intersection matrix describing overlaps between ReMap2020's ChIP-seq tracks
and \code{\link{enhancers_f5}}. To find overlapping regions enhancers were extended
by 500bp in both directions.
}
\keyword{datasets}
