% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remap_entrez}
\alias{remap_entrez}
\title{ReMap2020 gene level interaction matrix}
\format{
A Matrix with 23274 rows and 5728 columns. Row names corresponds to gene's
ENTREZ IDs, column names are formatted as ExperimentID.TranscriptionFactor.Biotype.
}
\usage{
remap_entrez
}
\description{
An matrix describing interactions between ReMap2020's ChIP-seq tracks
and human genes (hg38) as defined by \code{\link{promoters}}. The interaction score
for each gene and transcription factor is a sum of transcription factor occurences
in the gene promoter. Promoters were assigned to thier target genes based on ENTREZ
IDs.
}
\keyword{datasets}
