% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remap_meta}
\alias{remap_meta}
\title{ReMap2020 metadata}
\format{
A \code{data.table} with 5798 rows and 6 columns.
\describe{
\item{id}{Character giving internal experiment ID.}
\item{tf}{Character giving transcription factor name.}
\item{tf_dbd}{Character giving transcription factor DNA binding domain
family, as per CIS BP database.}
\item{biotype}{Character giving experiment biological origin.}
\item{study}{Character giving study ID.}
\item{condition}{Character specifiying experiment conditions or treatment.}
}
}
\usage{
remap_meta
}
\description{
Metadata associated with \code{remap_promoters} and \code{remap_enhancers}.
}
\keyword{datasets}
