% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remap_mini}
\alias{remap_mini}
\title{xcore example molecular signatures}
\format{
A \code{dgCMatrix} with 14191 rows and 600 columns holding
interaction matrix for subset of ReMap2020 molecular signatures against
FANTOM5 annotation. Rows corresponds to FANTOM5 promoters and columns to
signatures.
}
\usage{
data(remap_mini)
}
\description{
Molecular signatures data intended for use in xcore vignette and examples.
It is build ReMap2020 molecular signatures constructed against FANTOM5
annotation, which can be found in \code{xcoredata} package.
Here the data is only a subset limited to core promoters
(\code{promoters_f5_core}) and randomly selected 600 signatures.
}
\keyword{datasets}
