% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remap_promoters}
\alias{remap_promoters}
\title{ReMap2020 and FANTOM5 promoters intersection matrix}
\format{
A Matrix with 209911 rows and 5728 columns. Row names corresponds to promoters
names, column names are formatted as ExperimentID.TranscriptionFactor.Biotype.
}
\usage{
remap_promoters
}
\description{
An intersection matrix describing overlaps between ReMap2020's ChIP-seq tracks
and \code{\link{promoters}}. To find overlapping regions promoters were extended
by 500bp in both directions.
}
\keyword{datasets}
