% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ridgePValues}
\alias{ridgePValues}
\title{Calculate se and p-values for ridge regression betas}
\usage{
ridgePValues(x, y, beta, lambda, standardizex = TRUE, svdX = NULL)
}
\description{
Calculate se and p-values for ridge regression betas
}
