% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rinderpest_mini}
\alias{rinderpest_mini}
\title{xcore example expression data}
\format{
A \code{matrix} with 14191 rows and 6 columns holding expression
counts from CAGE-seq experiment. Rows corresponds to FANTOM5 promoters and
columns to time points at which expression was measured 0 and 24 hours post
infection.
}
\usage{
data(rinderpest_mini)
}
\description{
Expression data intended for use in xcore vignette and examples. It is build
from FANTOM5's 293SLAM rinderpest infection time course dataset. Here the
data is only a subset limited to core promoters (\code{promoters_f5_core}).
}
\keyword{datasets}
