% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsq}
\alias{rsq}
\title{Calculate $R^2$}
\usage{
rsq(y, yhat, offset)
}
\arguments{
\item{y}{numeric vector of observed expression values.}

\item{yhat}{numeric vector of predicted expression values.}

\item{offset}{numeric vector giving basal expression level.}
}
\value{
numeric vector
}
\description{
Calculate $R^2$
}
