% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{run_fgsea}
\alias{run_fgsea}
\title{Run fgsea() wrapper/convenience function}
\usage{
run_fgsea(ranks, pathways, nperm = 1)
}
\arguments{
\item{ranks}{a list of ranks as produced by ranks_from_FC_table()}

\item{pathways}{a list of signatures as produced by matrix_to_list()}

\item{nperm}{number of permutation done by fgsea(), default nperm = 1 to calculate the enriment scores without calculating p-value, you can calculate p-value by using nperm = 1000 or nperm = 10000 but it can be prohibitively slow}
}
\value{
a list, each element represent fgsea() result for one differential expression comparison, across all signatures
}
\description{
Run fgsea() wrapper/convenience function
}
