% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{select_top_mean_up}
\alias{select_top_mean_up}
\title{Select rows with highest mean}
\usage{
select_top_mean_up(matrix, n = 30)
}
\arguments{
\item{matrix}{a matrix or data.frame.}

\item{n}{a number of rows to return.}
}
\value{
a matrix or data.frame with \code{n} rows.
}
\description{
Select rows with highest mean
}
