% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsolete.R
\name{select_top_var}
\alias{select_top_var}
\title{Select top variable rows}
\usage{
select_top_var(matrix, n = 30)
}
\arguments{
\item{matrix}{a matrix or data.frame.}

\item{n}{a number of rows to return.}
}
\value{
a matrix or data.frame with \code{n} rows.
}
\description{
Select top variable rows
}
