% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectParams4dynamicTreeCut}
\alias{selectParams4dynamicTreeCut}
\title{Select cutHeight and deepSplit parameters for dynamicTreeCut}
\usage{
selectParams4dynamicTreeCut(
  dendro,
  distM,
  ref_cl,
  minClusterSize = 3,
  c_deepSplit = 1:4,
  c_cutHeight = seq(from = 0.5, to = 6, by = 0.2)
)
}
\description{
Select cutHeight and deepSplit parameters for dynamicTreeCut
}
