% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_genomic_regions.R
\name{simplify_gencode_annotation}
\alias{simplify_gencode_annotation}
\title{Simplify Gencode Annotation}
\usage{
simplify_gencode_annotation(annotation)
}
\arguments{
\item{annotation}{character vector with Gencode annotation that can be used to grep() labels as: "intron", "exon", "three_prime_UTR", "five_prime_UTR", "promoter"}
}
\value{
character vector with simplified labels
}
\description{
Simplify Gencode Annotation
}
