% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplifyInteractionMatrix}
\alias{simplifyInteractionMatrix}
\title{Simplify Interaction Matrix over groups}
\usage{
simplifyInteractionMatrix(mat, gr, method = "majority")
}
\arguments{
\item{mat}{dgCMatrix interaction matrix such as produced by
\code{\link{getInteractionMatrix}}.}

\item{gr}{Character vector specifying features groups. Must have length equal
to number of columns in \code{mat}.}

\item{method}{String. Currently only majority voting strategy is available.}
}
\value{
dgCMatrix
}
\description{
Simplify Interaction Matrix over groups
}
