% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stoufferZMethod}
\alias{stoufferZMethod}
\title{Combine Z-scores using Stouffer's method}
\usage{
stoufferZMethod(z)
}
\arguments{
\item{z}{a numeric vector of Z-score to combine.}
}
\value{
a number giving combined Z-score.
}
\description{
Stouffer's Z-score method is a meta-analysis technique used to combine the
results from independent statistical tests with the same hypothesis. It is
closely related to Fisher's method, but operates on Z-scores instead of
p-values
(\href{https://en.wikipedia.org/wiki/Fisher\%27s_method}{Wikipedia article}).
}
