% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsetWithMissing}
\alias{subsetWithMissing}
\title{Subset keeping missing}
\usage{
subsetWithMissing(mat, rows)
}
\arguments{
\item{mat}{matrix}

\item{rows}{character}
}
\value{
a matrix
}
\description{
Subset matrix keeping unmatched rows as NA.
}
