% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{summaryString}
\alias{summaryString}
\title{Object Summary as a string}
\usage{
summaryString(...)
}
\arguments{
\item{...}{arguments passed to \code{summary}.}
}
\value{
String.
}
\description{
Return object summary as a string that can be placed on a plot.
}
