% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tau}
\alias{tau}
\title{Calculate Tau}
\usage{
tau(x)
}
\arguments{
\item{x}{Numeric vector giving expression of one gene in different tissues.}
}
\value{
Tau tissue-specificity metric.
}
\description{
\code{tau} calculate Tau tissue-specificity metric.
\deqn{\tau = \frac{\sum_{i=1}^{n} (1-\hat{x_i})}{n-1}; \hat{x_i} = \frac{x_i}{max(x)}}
\link[https://academic.oup.com/bib/article/18/2/205/2562739]{Reference}.
}
