% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chip_atlas_meta}
\alias{chip_atlas_meta}
\title{ChIP-Atlas metadata}
\format{
A \code{data.frame} with 15133 rows and 11 columns.
\describe{
\item{id}{String specifying ExperimentID.}
\item{TF}{String specifying TranscriptionFactor.}}
\item{origin}{String specifying experiment biological origin.}
\item{cell}{String specifying cell line in which the experiment was conducted.}
\item{description}{String giving descriptions of primary tissue and tissue diganosis.}
\item{name}{String specifying feature name in format TranscriptionFactor_Origin_Cell_ExperimentID.}
\item{number_of_promoters_with_peak}{Integer giving number of promoters associated with the feature.}
\item{number_of_enhancers_with_peak}{Integer giving number of enhancers associated with the feature.}
\item{number_of_geneSymbols_with_peak}{Integer giving number of enhancers associated with the feature based on gene symbol.}
\item{number_of_geneEntrez_with_peak}{Integer giving number of genes associated with the feature based on ENTREZ ID.}
\item{total_number_of_peaks}{Integer giving total number of peaks associated with the feature.}
}
}
\usage{
chip_atlas_meta
}
\description{
ChIP-Atlas metadata... # TODO
}
\keyword{datasets}
