% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimateStat}
\alias{estimateStat}
\title{Estimate a goodness of fit stat}
\usage{
estimateStat(x, y, u, s, method = "cv", nfold = 10, statistic = R2, alpha = 0)
}
\description{
Estimate a goodness of fit stat
}
