% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getInteractionMatrix}
\alias{getInteractionMatrix}
\title{Compute interaction matrix}
\usage{
getInteractionMatrix(a, b, ext = 500, count = FALSE)
}
\arguments{
\item{a}{GRanges object.}

\item{b}{GRanges object.}

\item{ext}{Integer specifying number of base pairs the \code{a} coordinates
should be extended in upstream and downstream directions.}

\item{count}{Logical indicating if matrix should hold number of overlaps
between \code{a} and \code{b} or if FALSE presence / absence indicators.}
}
\value{
Sparse matrix of class dgCMatrix, with rows corresponding to
\code{a} and columns to \code{b}. Each cell holds a number indicating
how many times \code{a} and \code{b} overlapped.
}
\description{
\code{getInteractionMatrix} construct interaction matrix between two Granges
objects. Names of object \code{a} became row names and names of \code{b}
column names.
}
