% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotExperimentsHeatmap}
\alias{plotExperimentsHeatmap}
\title{Experiment's heatmap}
\usage{
plotExperimentsHeatmap(
  mat,
  meta,
  method = "pearson",
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  angle_col = 45,
  treeheight_col = 0,
  ...
)
}
\arguments{
\item{mat}{dgCMatrix interaction matrix such as produced by
\code{\link{getInteractionMatrix}}.}

\item{meta}{data.table metadata associated with \code{mat}. Should contain
\code{id} column, which is used to subset \code{mat}. Other character
columns are used as heatmap annotations.}

\item{method}{a character string indicating which method should be used to
calculate distances between \code{mat} columns. Supported values: "pearson",
"kendall", "spearman" (the actual distance is calculated as
(1 - cor) /2), "binary".}

\item{cluster_rows}{boolean values determining if rows should be clustered or \code{hclust} object,}

\item{cluster_cols}{boolean values determining if columns should be clustered or \code{hclust} object.}

\item{angle_col}{angle of the column labels, right now one can choose only from few 
predefined options (0, 45, 90, 270 and 315)}

\item{treeheight_col}{the height of a tree for columns, if these are clustered. 
Default value 50 points.}

\item{...}{arguments passed to \code{\link[pheatmap]{pheatmap}} function.}
}
\value{
Invisibly a pheatmap object. See \code{\link[pheatmap]{pheatmap}} for
more informations.
}
\description{
\code{plotExperimentsHeatmap} crate a heatmap for a group o experiments
(columns) based on interaction matrix. As function does not use sparsity
aware methods for distance calculation it is advised to not use too large
matrices. \code{\link[colorspace]{qualitative_hcl}} color palette is used
to allow higer number of categories in the annotation. No scaling is applied
to \code{mat}!
}
