% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{poolSE}
\alias{poolSE}
\title{Pool Standard Error / Standard Deviation}
\usage{
poolSE(x)
}
\arguments{
\item{x}{Numeric vector of standard errors to pool.}
}
\value{
Number giving pooled standard error.
}
\description{
Pooled standard error is calculated following (Cohen 1977) formulation for
pooled standard deviation.
TODO check out https://www.statisticshowto.com/find-pooled-sample-standard-error/, https://www.statisticshowto.com/pooled-standard-deviation/
}
