% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{runLinearRidge}
\alias{runLinearRidge}
\title{Linear ridge regression}
\usage{
runLinearRidge(x, y, offset, alpha = 0, standardize = TRUE, ...)
}
\arguments{
\item{x}{\code{x} matrix as in \code{glmnet}.}

\item{y}{response \code{y} as in \code{glmnet}.}

\item{offset}{Offset vector (matrix) as in \code{glmnet}}

\item{...}{Other arguments that can be passed to \code{glmnet}}
}
\value{
an object of class "cv.glmnet" is returned. See
\code{\link[glmnet]{cv.glmnet}} for more details.
}
\description{
Wrapper around \code{\link[glmnet]{cv.glmnet}} to run linear ridge regression
with lambda selection using cross-validation.
}
