% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplifyInteractionMatrix}
\alias{simplifyInteractionMatrix}
\title{Simplify Interaction Matrix}
\usage{
simplifyInteractionMatrix(mat, alpha = 0.5, colname = NA)
}
\arguments{
\item{mat}{dgCMatrix interaction matrix such as produced by
\code{\link{getInteractionMatrix}}.}

\item{alpha}{Number between 0 and 1 specifying voting threshold. Eg. for 3
column matrix alpha 0.5 will give voting criteria >= 2.}

\item{colname}{character giving new column name.}
}
\value{
dgCMatrix
}
\description{
Simplify Interaction Matrix
}
