% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translateCounts}
\alias{translateCounts}
\title{Translate counts matrix rownames}
\usage{
translateCounts(counts, dict)
}
\arguments{
\item{counts}{matrix of expression values.}

\item{dict}{named character vector mapping \code{counts} rownames to new
values.}
}
\value{
matrix of expression values with new rownames.
}
\description{
\code{translateCounts} renames counts matrix rownames according to supplied
\code{dict}ionary. Function can handle many to one assignments by summing
over \code{counts} rows. Other types of ambiguous assignments are not
supported.
}
\examples{
#TODO also add tests

}
